+++ bfabric/docs/changelog.md
@@ -25,6 +25,7 @@ We will continue to track the B-Fabric version in the `bfabric-scripts` package 
 - `bfabric.entities.core.entity_reader` which allows reading entities by URI, ID, and general queries.
 - `Entity.refs` resolves all references of entities generically, writes the data into data_dict and is compatible with pre-loaded references from B-Fabric.
 - `Entity.uri` property to get the URI of an entity.
+- `Entity.custom_attributes` and `Entity.has_custom_attributes` for reading custom attribute values.
 - `bfabric.entities.cache` which supersedes `bfabric.experimental.cache` (temporarily kept in tree).
 - `HasOne` and `HasMany` should be able to use resolved entities when loading with `fulldetails=True`.